v1.0.5.6 (M+D) working deployment

+++ client/src/components/OrderList.js
@@ -49,45 +49,53 @@ export default class ItemList extends Component {
             })
     }
 
-    handleRemoveOrder(id) {
+    handleRemoveOrder(id, name) {
         console.log('remove this id:', id)
+        console.log('remove this name:', name)
         // const change = this.state.approves.filter(item => item._id !== id)
-
-        // DELETE Request
-        axios.delete(`/orders/${id}`, {
-            headers: {
-                'x-auth': localStorage.getItem('token')
-            }
-        })
-            .then((response) => {
-                console.log('response data', response.data)
-                this.setState((prevState) => ({
-                    approves: prevState.approves.filter(item => item._id !== response.data._id)
-                }))
-                // remove if not required
-                this.setState((prevState) => ({
-                    confirms: prevState.confirms.filter(item => item._id !== response.data._id)
-                }))
-
-                this.setState((prevState) => ({
-                    completed: prevState.completed.filter(item => item._id !== response.data._id)
-                }))
-
-                // var NewItems = this.props.items.filter(item => {
-                //     return item !== itemToDelete
-                // })
-
-                // this.setState({
-                //     items: NewItems
-                // })
-                // console.log("items after []:", NewItems)
-                // console.log("item deleted :", this.props.items.splice(this.props.items.indexOf(itemToDelete), 1))
+        // add confirmation here
+        if (window.confirm(`are you sure you want to delete the order of ${name}`)) {
+            console.log('delete here')
 
 
+            // DELETE Request
+            axios.delete(`/orders/${id}`, {
+                headers: {
+                    'x-auth': localStorage.getItem('token')
+                }
             })
-            .catch((err) => {
-                console.log(err)
-            })
+                .then((response) => {
+                    console.log('response data', response.data)
+                    this.setState((prevState) => ({
+                        approves: prevState.approves.filter(item => item._id !== response.data._id)
+                    }))
+                    // remove if not required
+                    this.setState((prevState) => ({
+                        confirms: prevState.confirms.filter(item => item._id !== response.data._id)
+                    }))
+
+                    this.setState((prevState) => ({
+                        completed: prevState.completed.filter(item => item._id !== response.data._id)
+                    }))
+
+                    // var NewItems = this.props.items.filter(item => {
+                    //     return item !== itemToDelete
+                    // })
+
+                    // this.setState({
+                    //     items: NewItems
+                    // })
+                    // console.log("items after []:", NewItems)
+                    // console.log("item deleted :", this.props.items.splice(this.props.items.indexOf(itemToDelete), 1))
+
+                })
+                .catch((err) => {
+                    console.log(err)
+                })
+
+        } else { console.log('dont delete the order') }
+
+
 
         // this.setState(
         //     prevState => ({
@@ -268,7 +276,7 @@ export default class ItemList extends Component {
                                                 <button>Update</button>
 
                                                 <button onClick={() => {
-                                                    this.handleRemoveOrder(item._id)
+                                                    this.handleRemoveOrder(item._id, item.customer.fullName)
                                                 }}>Remove</button>
                                                 <button onClick={() => {
                                                     this.handleApproveOrder(item._id)
@@ -290,7 +298,7 @@ export default class ItemList extends Component {
                     <input placeholder="Search Order" />
                     <button>Search</button>
 
-                    <button>Add new Order</button>
+                    <Link to='/menu'><button>Add new Order</button></Link>
                     <table className="listing-table" >
                         <thead className="listing-table" >
                             <tr>
@@ -314,7 +322,7 @@ export default class ItemList extends Component {
                                                 item.customer.eventDate.substr(8, 2) + "/" + item.customer.eventDate.substr(5, 2) + "/" + item.customer.eventDate.substr(0, 4)
                                             }</td>
                                             <td className="listing-table" ><button onClick={() => {
-                                                this.handleRemoveOrder(item._id)
+                                                this.handleRemoveOrder(item._id, item.customer.fullName)
                                             }}>Remove</button></td>
                                             <td className="listing-table" ><button onClick={() => {
                                                 this.handleCompleteOrder(item._id)
@@ -341,7 +349,7 @@ export default class ItemList extends Component {
                                 <td className="listing-table">Remove</td>
                             </tr>
                         </thead>
-                        
+
                         <tbody>
                             {
                                 this.state.completed.map((item, i) => {
@@ -353,7 +361,7 @@ export default class ItemList extends Component {
                                                 item.customer.eventDate.substr(8, 2) + "/" + item.customer.eventDate.substr(5, 2) + "/" + item.customer.eventDate.substr(0, 4)
                                             }</td>
                                             <td className="listing-table" ><button onClick={() => {
-                                                this.handleRemoveOrder(item._id)
+                                                this.handleRemoveOrder(item._id, item.customer.fullName)
                                             }}>Remove</button></td>
                                         </tr>
                                     )
